feat : integration user route section with real apI
